Create select even&odd position components in a string
"Hello, world!"[1:9:2] → This gives "el,w".it starts with the character at index 1, and goes until (but not including) the character with index 9. Here, though, it includes every other character. The :2 at the end forces it to only take characters 1, 3, 5, and 7.
"Hello, world!" [::3] → This gives "Hl r!". Leaving the first two spots blank tells it to look at the entire string, but putting :3 at the end says to only take every third character. It does so starting at the beginning, and takes the characters at indices 0, 3, 6, 9, and 12 (H, l, space, r, and !).